make estimator.cpp smaller
